exercise

+++ src/hust/soict/dsai/aims/media/Disc.java
@@ -4,7 +4,7 @@ public class Disc extends Media {
     protected int length;
     protected String director;
     public Disc(){};
-    public Disc(float length, String director) {
+    public Disc(int length, String director) {
         super();
         this.length = length;
         this.director = director;
